requirements.txt
defusedxml==0.4.1
Django==1.7.2
django-appconf==1.0.1
django-haystack==2.3.1
django-imagekit==3.2.6
django-registration-redux==1.1
mysql-connector-python==2.0.2
oauthlib==0.7.2
pbr==0.10.7
pilkit==1.1.12
Pillow==2.7.0
Pinax==0.9a2
python3-openid==3.0.5
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
stevedore==1.2.0
virtualenv==12.0.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.2
Whoosh==2.6.0
